Backend/requirements.txt
django
djangorestframework
django-cors-headers
dj-database-url
psycopg2-binary
python-dotenv
